Code spins up motor and correctly reads current and bus voltages on a cycle by cycle basis at 36kHz. There is a hall sensor timing issue that limits the possible speed of the motor; the electrical field lags or leads the motor field, which results in a large current buildup towards the start or end of the cycle. This really need fixing in software, since it it likely that any motor will have slightly offset hall sensors.

+++ Core/Src/stm32f3xx_it.c
@@ -180,7 +180,7 @@ void DMA1_Channel1_IRQHandler(void)
   HAL_DMA_IRQHandler(&hdma_adc1);
   /* USER CODE BEGIN DMA1_Channel1_IRQn 1 */
 	BLDCCommuteHall();
-	extern uint32_t adcBuff1[3];
+	extern float adcBuff1[3];
 	adcBuff1[0]=(float)measurement_buffers.RawADC[0][0];
 	adcBuff1[1]=(float)measurement_buffers.RawADC[0][1];
 	adcBuff1[2]=(float)measurement_buffers.RawADC[0][2];
